refactor: simplify filters by moving back to picomatch
