+++ src/index.ts
@@ -6,5 +6,4 @@ export { StatsFileSource } from './data/StatsFileSource';
 
 export { AtlasError, AtlasValidationError } from './utils/errors';
 export { createAtlasMiddleware } from './utils/middleware';
-export { fuzzyFilterModules } from './utils/search';
 export { createStatsFile, validateStatsFile, getStatsMetdata, getStatsPath } from './utils/stats';
